Moved features to FeatHub. Added to parts list.
Removed the feature list and descriptions from the README.md to the descriptions of FeatHub. This will allow users to also vote and suggest on additional features that I haven't thought of. Added a line item in the parts list for the camera and mic so I can do facial recognition and voice controls.